move TrackingOptIn to its own class for easier testing
